pyproject.toml
[project]
name = "nbwf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.0.1",
    "faker>=30.3.0",
    "fastparquet>=2024.5.0",
    "flytekit",
    "flytekitplugins-envd>=1.13.8",
    "flytekitplugins-spark>=1.13.8",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.2",
    "nbformat>=5.10.4",
    "pandas==2.2.2",
    "pip>=24.2",
    "plotly>=5.24.1",
    "polars>=1.9.0",
    "pyarrow>=17.0.0",
    "scikit-learn==1.5.1",
    "union>=0.0.80",
]

[tool.uv.sources]
flytekit = { git = "https://github.com/flyteorg/flytekit.git", rev = "60499483afd43585279066389357dfe7cfb544e5" }
